requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.14.3
boto3==1.42.59
botocore==1.42.59
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cloudscraper==1.2.71
cryptography==46.0.5
filelock==3.25.0
idna==3.11
jmespath==1.1.0
kenpompy==0.5.0
lxml==6.0.2
MechanicalSoup==1.4.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
platformdirs==4.9.2
pyarrow==23.0.1
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==4.3.0
sortedcontainers==2.4.0
soupsieve==2.8.3
tomlkit==0.14.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
